LU-3397 lprocfs: create "export" /proc file on server
Similar to the "import" file on the client for each client-to-server
connection, it would be useful to have a file on the server in the
per-client directory obdfilter/*/exports/$NID/export. This would
contain export connection information as in the "import" file.

Also, this patch remove useless function lprocfs_exp_rd_nid().

Signed-off-by: Liu Ying <emoly.liu@intel.com>
Change-Id: I60896090e3a8ad872141a8d4299f0698f0a5636a