LU-1993 llite: debugging for ll_file_open LASSERT
Add debugging for LASSERTF(it_disposition(it, DISP_ENQ_OPEN_REF)
in ll_file_open(), since this is a rarely hit failure under racer,
and it would be useful to get more information if this is hit
again.  Print the full intent disposition, as well as the status,
in case Oleg's earlier comment about bailing out on any error is
actually the case here.

Test-Parameters: envdefinitions=DURATION=1800 testlist=racer
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Ie03f706bcb32b03d94c84c5d80dae197813ebbe5