LU-3270 statahead: NOT free entry name with RPC in flight
There is a race in the code where statahead thread would send an
RPC and then exit without waiting for RPC to complete. On thread
exit all the statahead entries (in spite of replied or not) will
be freed, including the entry name which is allocated inside the
statahead entry. On the other hand, the entry name is referenced
by related statahead RPC, so accessing the entry name in the RPC
context after the statahead entry freed is invalid.

To avoid that, the new mechanism will allocate entry name in the
'minfo' which is referenced by the statahead RPC and will NOT be
freed when the statahead entry is freed before the statahead RPC
replied. The RPC callback handler can free the orphan entry name
together with the 'minfo'. On the other hand, If there is no RPC
sent for the statahead entry (not necessary), then allocate/free
the entry name together the statahead entry.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Iec3a011f44b8ede8d6a133a07edf023008a3145e