LU-3338 llite: Fix IOC_MDC_GETFILESTRIPE vmalloc thrashing
When allocating a reply buffer for the striping information don't
default to assuming the unlikely worst case.  Instead, assume the
common case and size the buffer based on the default stripe count.
This has a number of advantages.

* For large filesystems where the worst case striping information
  may exceed 16k it converts a vmalloc(16k+) to a small kmalloc().
  This prevents the client from thrashing on the vmalloc() spin
  lock when attempting to allocate the large reply buffers.

* A reply buffer of exactly the right size (no larger) is allocated
  in the overflow case.  These larger reply buffers are still
  unlikely to exceed the 16k limit where a vmalloc() will occur.

* Saves memory in the common case.

The reason this patch works is because the ptlrpc layer is smart
enough to reallocate the reply buffer when an overflow occurs.
Therefore the client doesn't have to drop the incoming reply and
send a new request with a larger reply buffer.

It's also worth mentioning that the reply buffer always contains
a significant amount of extra padding.  This means that even
files striped wider than the default have a good chance of fitting
in the allocated reply buffer.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Change-Id: Ic6bab2aca208db1f4abcd87974a5879d06d0f2f1