LU-3270 statahead: avoid freeing minfo with request in flight
There is a race in the code where statahead thread would send an
RPC and then exit without waiting for RPC to complete.
On thread exit minfo attached to sai would be freed, but it's also
referenced from RPC to be used when RPC reply is received.
As such this patch stops freeing minfo from ll_sai_cleanup
and instead minfo is now freed either in ll_statahead_post after RPC
is received for normal operations, or in ll_statahead_interpret if we
received a reply after statahead thread terminated.

Change-Id: Ib23af5d3f009a0768d119522382fc863c449f9b4
Signed-off-by: Oleg Drokin <oleg.drokin@intel.com>