LU-3346 obt: Completely remove the usage of obt_sb
The assignment of obt->obt_sb was removed in a recent project.
The peace of code with obt->obt_sb usage from this function has
never executed because of smart compiler remove it. This BUG
was not discovered because we have IR so that even the timeout
is much short it can still survive.

Signed-off-by: Dmitry Eremin <dmitry.eremin@intel.com>
Change-Id: I4b87e887e4f8799814883ca0a58193446c418739