LU-3403 llite: error of listxattr when buffer is small
According to the standard, listxattr(2) should return -1
and errno should be set to ERANGE if the size of the list
buffer is too small to hold the result. However ll_listxattr()
will return a value bigger than the size of buffer in some cases.

Let's assume listxattr(2) returns SIZE when it is called with a
large enough list buffer. If it's called with a list buffer whose
size is smaller than SIZE but bigger than (SIZE - 12), then
listxattr(2) will return SIZE too. This patch fixed the problem.

Signed-off-by: Li Xi <pkuelelixi@gmail.com>
Change-Id: Ifb29ae900a5ee57dfde8781f7a008b6129c7bf91