LU-3446 mdd: changelog index to survive MDT restart
If all changelog records are cleared the current changelog index is
initialized to zero when the MDT starts. This confuses applications
that rely on the index to track their position in the changelog
history. To avoid this, initialize the current index from the
registered user with the highest end record index.

Add a sanity.sh regression test. Update changelog test function
test_160() and its helpers according to current style guidelines.

Lustre-commit: 8a4fa4e5c041f94e6dffa5524542e97b5c5f71fd
Lustre-change: http://review.whamcloud.com/6642

Signed-off-by: Ned Bass <bass6@llnl.gov>
Change-Id: I24fdf50b80904ef58305f98431f171dc7f7b5e5b