LU-1669 vvp: Use lockless __generic_file_aio_write
Testing multi-threaded single shard file write performance has shown the
inode mutex to be a limiting factor when using the generic_file_aio_write
function. To work around this bottle neck, this change replaces the
locked version of that call with the lock less version, specifically,
__generic_file_aio_write.

In order to maintain posix consistency, Lustre must now employ it's own
locking mechanism in the higher layers. Currently write are protected
using the lli_write_mutex in the ll_inode_info structure. Since truncate
operations are protected using the lli_trunc_sem, we need to down that
semaphore during writes to prevent simultaneous conflicting write and
truncate operations.

In order to garner any real performance benefits from this change, the
lli_write_mutex needs to be broken up into a range lock, to allow for
parallel write operations to unique regions of a file. This work is left
to be done in a follow up patch.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: I0023132b5d941b3304f39f015f95106542998072