LU-2902 test: ost stat parameters were not present
A few tests use roc_hit to sort out cache_hits. These tests
were failing for a while because the lproc stats needed were not
present.

Use cache_access to identify the number of OST stat
producers avaible and error if this number is not OSTCOUNT

Awk/Bash clarification implmented.

Signed-off-by: Keith Mannthey <keith.mannthey@intel.com>
Change-Id: Iea501433ddcdb367fe5d9614361b4aff049cb1f1