LU-3335 scrub: OI scrub on OST
OI scrub should has the ability to handle kinds of OI, including
both the OI files on MDT and the /O directory on OST.

We trust the FID in LMA for both MDT objects and OST objects. So
if some /O sub-item does not match related LMA, then the /O will
be updated, instead of the LMA.

To guarantee that the OI scrub can run without MDT0 involved for
FLDB, the OST object needs to store some flag in its LMA to tell
OI scrub that it is for an OST object, no need to query the MDT0.

Lustre-commit: I2da3816b28a7c6d5ad0b7f8d5b43f98a9886ff0c
Lustre-change: http://review.whamcloud.com/6669

Test-Parameters: testlist=sanity-scrub
Signed-off-by: Fan Yong <fan.yong@intel.com>
Signed-off-by: Lai Siyao <lai.siyao@intel.com>
Change-Id: Ie126dc83f9464d872ef40407b7f93ebc44564c64