LU-2264 test: Create cpu load for testing
It can be useful to add extra cpu pressure for testing.
Client nodes are likely to use extra resources.

Via processor.sh we can add cpu load. It has a dumb bash loop
that is spaws by defaut 2 threads per cpu. It is adjustable
as to the number of threads per cpu via the -m option.

This is nothing fancy just a way to introduce non IO cpu load.

Perhaps we could add a -j to do a cpu power state change mode
to interduce jitter to the system.

Signed-off-by: Keith Mannthey <keith.mannthey@intel.com>
Change-Id: I18a162f52eadc412110fe61d02c1ceaec7f99297