LU-2919 hsm: Implementation of exclusive open
In this patch, a framework of lease is implemented. However,
only exclusive lease is supported right now.

To apply a lease, MDS_OPEN_LEASE must be set to open the file,
open lock is returned to the client side to hold a lease. From now
on, if this file is opened with a conflicted mode, the open lock
will be revoked so the client who holds the lease knows the lease
is broken.

To release a lease, normal close is used. The client will revoke
the open lock before sending CLOSE request.

Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Change-Id: I181edbb0b31fcf60858656170528d2ffda9c9bb6