LU-2264 test: Create memory load for testing
It can be useful to add extra memory pressure for testing.
Client nodes are likely to use extra resources.

rewrite memhog.c to include an automatic mode. ./memhog -a will
consume all but 160-128mb of Ram and Swap. The mode consumes
then exits at this time. Original behavior is maintained with
the exception of moving the chunk size to 32mb (down from 128m)

Signed-off-by: Keith Mannthey <keith.mannthey@intel.com>
Change-Id: I590b08479c559a846fee166d2c8ea28cbe6a89e4