LU-3494 libcfs: Add relocation function to libcfs heap
When changing the values of the node in the heap which might affects
its ranking, we have to remove the node from the heap, change the
values and then insert the node into the heap again. The process of
inserting into the heap might fails because of memory insufficiency.
Sometimes, it is not convienient and efficient for a caller to handle
such kind of failure.

This patch adds a new function to do the job.

Signed-off-by: Li Xi <pkuelelixi@gmail.com>
Change-Id: I31ce530d750341f6479470ba61457d9ed03afed3