LU-3486 llite: remove ll_d_root_ops
Mnt root dentry never calls lookup/revalidate/..., so it's redundant
to set dentry_operations on it. But its attribute may be invalid
before access, this won't cause any issue because it always exists,
and the only operation depends on its attribute is .permission,
which has revalidated it in lustre code.

So it's okay to remove ll_d_root_ops, and remove unnecessary checks
in lookup/revalidate/statahead.

This patch also includes several small fixes:
* don't d_add() for create only files, because the dentry is fake,
  and will be release right after use.
* alloc ll_dentry_data may fail, add check for this.

Signed-off-by: Lai Siyao <lai.siyao@intel.com>
Change-Id: I97b71a9540146809171d80aa5d1a9d4f308faf7f