LU-1346 libcfs: add libcfs_cleanup makefile target
Add a "libcfs_cleanup" makefile target so that it is easier for
developers to run the contrib/scripts/libcfs_cleanup.sed script
on the source files after some cleanup has been done.  This is
not run automatically, only upon request.

Exclude the gnilnd subdirectory, as James prefers to update this
himself.

Fix up a few nits in the libcfs_cleanup.sed script.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: If61c4318e9d4e7bb59592dfef5c9ee4f973ebbe5