LU-1267 lfsck: enhance RPCs (2) for MDT-OST consistency
The LFSCK on the OST uses OST_SET_INFO RPC to notify the LFSCK on the
MDT about the LFSCK progress for the layout consistency verification.
And uses OST_GET_INFO RPC to query the LFSCK status on the MDT.

The LFSCK RPC from OST to MDT is sent via the reserse import. It may
be changed to normal import when we handle orphan OST-objects in the
future.

OST/OFD supports incoming obd_ops::{o_set_info_async,o_get_info}. We
do not want to make logic to be confused. So LFSCK RPC processing on
OST is implemented inside the LFSCK, and transparent to others.

Test-Parameters: testlist=sanity-lfsck
Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I138fa9b9ad8ab539379f25bb59ec04a1a482fddb