LU-3544 nfs: writing to new files will return ENOENT
This happend with SLES11SP2 Lustre client, which in turn acts as an
NFS server, exporting a subtree of an Lustre fs through NFS.

We detected that whenever we are writing to a new file using, for
e.g., 'echo blah > newfile', it will return ENOENT error. We found
out that this was caused by the anonymous dentry. In SLESS11SP2,
anonymous dentries are assigned '/' as the name, instead of an
empty string. When MDT handles the intent_open call, it will look
up the obj by the name if it is not an empty string, and thus
couldn't find it.

The solution is to distinguish anonymous dentries with regular
dentries on the client side. For anonymous ones, pass in empty
string so that MDT will look up by fid only.

Signed-off-by: Cheng Shao <cheng_shao@xyratex.com>
Change-Id: Ia8bd6f2814d05350d0a197df8a3ffd9729e2081b